Enabled sorting on heatmap in gene page Gene page redirects to search results heatmap view if gene query hits multiple genes
git-svn-id: svn://bar.ebi.ac.uk/trunk/atlas@7039 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-web/src/main/java/ae3/model/AtlasGene.java
@@ -186,27 +186,27 @@ public class AtlasGene {
 	
 
 	public int getCount_up(String ef, String efv){
-		return nullzero((Integer)geneSolrDocument.getFieldValue("cnt_efv_"+ef+"_"+EfvTree.encodeEfv(efv.toString()+"_up")));
+		return nullzero((Short)geneSolrDocument.getFieldValue("cnt_efv_"+ef+"_"+EfvTree.encodeEfv(efv.toString()+"_up")));
 	}
 
 	public int getCount_dn(String ef, String efv){
-		return nullzero((Integer)geneSolrDocument.getFieldValue("cnt_efv_"+ef+"_"+EfvTree.encodeEfv(efv.toString()+"_dn")));
+		return nullzero((Short)geneSolrDocument.getFieldValue("cnt_efv_"+ef+"_"+EfvTree.encodeEfv(efv.toString()+"_dn")));
 	}
 
 	public double getAvg_up(String ef, String efv){
-		return nullzero((Double)geneSolrDocument.getFieldValue("minpval_efv_" + ef+"_"+EfvTree.encodeEfv(efv.toString() + "_up")));
+		return nullzero((Float)geneSolrDocument.getFieldValue("minpval_efv_" + ef+"_"+EfvTree.encodeEfv(efv.toString() + "_up")));
 	}
 
 	public double getAvg_dn(String ef, String efv){
-		return nullzero((Double)geneSolrDocument.getFieldValue("minpval_efv_" + ef+"_"+EfvTree.encodeEfv(efv.toString() + "_dn")));
+		return nullzero((Float)geneSolrDocument.getFieldValue("minpval_efv_" + ef+"_"+EfvTree.encodeEfv(efv.toString() + "_dn")));
 	}
 
-	private static int nullzero(Integer i)
+	private static int nullzero(Short i)
 	{
 		return i == null ? 0 : i;
 	}
 
-	private static double nullzero(Double d)
+	private static double nullzero(Float d)
 	{
 		return d == null ? 0.0d : d;
 	}
